% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp}
\alias{disp}
\title{disp() Function}
\usage{
disp(Stages, SexDep, prob_dispersal_0, prob_dispersal_1, Distances)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
disp()
}
\keyword{Add}
