% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emig_prob.R
\name{emig_prob}
\alias{emig_prob}
\title{emig_prob is a function that is required by sp_dispersal.}
\usage{
emig_prob(Stages, prob_dispersal_0, prob_dispersal_1)
}
\description{
Note that this function ignores the SexDep argument.
}
