% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct.prob.R
\name{predextinct.prob}
\alias{predextinct.prob}
\title{predextinct.prob() Function}
\usage{
predextinct.prob(s, dirpath)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
predextinct.prob()
}
\keyword{Add}
