% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct.time.R
\name{predextinct.time}
\alias{predextinct.time}
\title{predextinct.time() Function}
\usage{
predextinct.time(s)
}
\arguments{
\item{s}{: The output of setscenario()}
}
\description{
This function predict the time of species extinction. Read the RangeshiftR output file from RangeShiftR::readPop.
}
\examples{
predextinct.time()
}
\keyword{predextinct.time}
