% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct_plot.R
\name{predextinct_plot}
\alias{predextinct_plot}
\title{x`predextinct_plot() Function}
\usage{
predextinct_plot(s, folder = "data/")
}
\arguments{
\item{s}{: The output of setscenario()}

\item{folder:}{\link{complete}.}
}
\description{
This function plot the Abundance and Occupancy of species.
This function run RangeshiftR output file from RangeShiftR::plotAbundance and RangeShiftR::plotOccupancy.
}
\examples{
predextinct_plot()
}
\keyword{predextinct_plot}
