% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct_pop.R
\name{predextinct_pop}
\alias{predextinct_pop}
\title{predextinct_time() Function}
\usage{
predextinct_pop(s, folder = "data/")
}
\arguments{
\item{s}{: The output of setscenario()}

\item{folder:}{\link{complete}.}
}
\description{
This function predict the population per 10 years for the species to experience extinction.
Read the RangeshiftR output file from RangeShiftR::readPop.
}
\examples{
predextinct_pop()
}
\keyword{predextinct_pop}
