% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct_prob.R
\name{predextinct_prob}
\alias{predextinct_prob}
\title{predextinct_prob() Function}
\usage{
predextinct_prob(s, folder = "data/")
}
\arguments{
\item{s}{: The output of setscenario()}

\item{folder:}{\link{complete}.}
}
\description{
This function predict the probability of species extinction. Read the RangeshiftR output file from RangeShiftR::readPop.
This function will estimate the extinction probability after 50 years.
}
\examples{
predextinct_prob()
}
\keyword{predextinct_prob}
