% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct_time.R
\name{predextinct_time}
\alias{predextinct_time}
\title{predextinct_time() Function}
\usage{
predextinct_time(s, folder = "data/")
}
\arguments{
\item{s}{: The output of setscenario()}

\item{folder:}{\link{complete}.}
}
\description{
This function predict the time (in years) for the species to experience extinction.
Read the RangeshiftR output file from RangeShiftR::readPop.
}
\examples{
predextinct_time()
}
\keyword{predextinct_time}
