% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setscenario.R
\name{setscenario}
\alias{setscenario}
\title{setscenario() Function}
\usage{
setscenario(csv_file)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
setscenario()
}
\keyword{Add}
