% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_stage.R
\name{sp_stage}
\alias{sp_stage}
\title{sp_stage is a function that is required by demo.}
\usage{
sp_stage(Stages, MaxAge, prob_reproduction, num_offsprings, prob_surv)
}
\description{
sp_stage is a function that is required by demo.
}
