% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_asc.R
\name{write_asc}
\alias{write_asc}
\title{write_asc() Function}
\usage{
write_asc(csv_file, folder = "data/")
}
\arguments{
\item{csv_file}{: The name of csv format file. It should be located in the working directory.}
}
\description{
This function is working for creating ASCII Raster format files of species from GBIF datasets.
}
\examples{
write_asc()
}
\keyword{csv_file}
