% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_demography.R
\name{demo}
\alias{demo}
\title{sp_demography() Function}
\usage{
demo(Stages, MaxAge, prob_reproduction, num_offsprings, prob_surv)
}
\arguments{
\item{Stages}{: Number of life-Stages (integer). Requires a minimum of 2, for "juvenile" (stage 0) and "adult". Maximum is 3.}

\item{MaxAge}{: Maximum age in years}

\item{prob_reproduction}{: Probability to produce offsprings}

\item{num_offsprings}{: Expected the number of offsprings}

\item{prob_surv}{: Probability to survive}
}
\description{
This function is a sub-function of setscenario().
}
\examples{
dontrun{
sp_demography()
}
}
\keyword{sp_demography}
