% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{demo() Function}
\usage{
demo(
  Stages,
  MaxAge,
  ReproductionType,
  prob_reproduction,
  num_offsprings,
  prob_surv
)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
demo()
}
\keyword{Add}
