% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp}
\alias{disp}
\title{disp() Function}
\usage{
disp(Stages, SexDep, prob_dispersal_0, prob_dispersal_1, Distances)
}
\arguments{
\item{Stages}{: Number of life-Stages (integer). Requires a minimum of 2, for "juvenile" (stage 0) and "adult". Maximum is 3.}

\item{SexDep}{: Sex-dependent emigration probability? (default: FALSE)}

\item{prob_dispersal_0}{: Dispersal probability (emigration probability) of stage 0 (juvenile)}

\item{prob_dispersal_1}{: Dispersal probability (emigration probability) of stage 1 (sub-juvenile)}

\item{Distnaces}{: Distances of species dispersal}
}
\description{
This function is a sub-function of setscenario().
}
\examples{
disp()
}
\keyword{disp}
