% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dispersal.R
\name{disp}
\alias{disp}
\title{sp_dispersal() Function}
\usage{
disp(Distances, Stages, prob_dispersal_0, prob_dispersal_1)
}
\arguments{
\item{Distances}{: Distances of species dispersal}

\item{Stages}{: Number of life-Stages (integer). Requires a minimum of 2, for "juvenile" (stage 0) and "adult". Maximum is 3.}

\item{prob_dispersal_0}{: Dispersal probability (emigration probability) of stage 0 (juvenile)}

\item{prob_dispersal_1}{: Dispersal probability (emigration probability) of stage 1 (sub-juvenile)}
}
\description{
This function is a sub-function of setscenario().
}
\examples{
dontrun{
sp_dispersal()
}
}
\keyword{sp_dispersal}
