% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct.prob.R
\name{predextinct.prob}
\alias{predextinct.prob}
\title{predextinct.prob() Function}
\usage{
predextinct.prob(s)
}
\arguments{
\item{s}{: The output of setscenario()}
}
\description{
This function predict the probability of species extinction. Read the RangeshiftR output file from RangeShiftR::readPop.
}
\examples{
predextinct.prob()
}
\keyword{predextinct.prob}
