% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predextinct.time.R
\name{predextinct.time}
\alias{predextinct.time}
\title{predextinct.time() Function}
\usage{
predextinct.time(s, dirpath)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
predextinct.time()
}
\keyword{Add}
