% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscenario.R
\name{runscenario}
\alias{runscenario}
\title{runscenario() Function}
\usage{
runscenario(s_list, dirpath)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function
}
\examples{
runscenario()
}
\keyword{Add}
