% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setscenario.R
\name{setscenario}
\alias{setscenario}
\title{setscenario() Function}
\usage{
setscenario(csv_file, folder = "data/")
}
\arguments{
\item{csv_file}{: The name of csv format file. It should be loacated in the working directory.}

\item{folder:}{\link{complete}.}
}
\description{
This function is setting up a parameters that can be used in RangeShiftR::RunRS() to run a simulation.
}
\examples{
setscenario()
}
\keyword{setscenario}
